+++ src/mpmToolbox/supplementary/FileDrop.java
@@ -9,15 +9,15 @@ import java.io.Reader;
 
 /**
  * This class makes it easy to drag and drop files from the operating
- * system to a Java program. Any <tt>java.awt.Component</tt> can be
- * dropped onto, but only <tt>javax.swing.JComponent</tt>s will indicate
+ * system to a Java program. Any java.awt.Component can be
+ * dropped onto, but only javax.swing.JComponents will indicate
  * the drop event with a changed border.
- * <p/>
- * To use this class, construct a new <tt>FileDrop</tt> by passing
- * it the target component and a <tt>Listener</tt> to receive notification
+ *
+ * To use this class, construct a new FileDrop by passing
+ * it the target component and a Listener to receive notification
  * when file(s) have been dropped. Here is an example:
- * <p/>
- * <code><pre>
+ *
+ *
  *      JPanel myPanel = new JPanel();
  *      new FileDrop( myPanel, new FileDrop.Listener()
  *      {   public void filesDropped( java.io.File[] files )
@@ -26,21 +26,21 @@ import java.io.Reader;
  *              ...
  *          }   // end filesDropped
  *      }); // end FileDrop.Listener
- * </pre></code>
- * <p/>
- * You can specify the border that will appear when files are being dragged by
- * calling the constructor with a <tt>javax.swing.border.Border</tt>. Only
- * <tt>JComponent</tt>s will show any indication with a border.
- * <p/>
- * You can turn on some debugging features by passing a <tt>PrintStream</tt>
- * object (such as <tt>System.out</tt>) into the full constructor. A <tt>null</tt>
- * value will result in no extra debugging information being output.
- * <p/>
  *
- * <p>I'm releasing this code into the Public Domain. Enjoy.
- * </p>
- * <p>Original author: Robert Harder, rob@iharder.net</p>
- * <p>Additional support: </p>
+ *
+ * You can specify the border that will appear when files are being dragged by
+ * calling the constructor with a javax.swing.border.Border. Only
+ * JComponents will show any indication with a border.
+ *
+ * You can turn on some debugging features by passing a PrintStream
+ * object (such as System.out) into the full constructor. A null
+ * value will result in no extra debugging information being output.
+ *
+ *
+ * I'm releasing this code into the Public Domain. Enjoy.
+ * 
+ * Original author: Robert Harder, rob@iharder.net
+ * Additional support: 
  * <ul>
  *  <li>September 2007, Nathan Blomquist -- Linux (KDE/Gnome) support added.</li>
  *  <li>December 2010, Joshua Gerth</li>
@@ -69,7 +69,7 @@ public class FileDrop
      * the top level container will change borders.
      *
      * @param c Component on which files will be dropped.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -87,12 +87,12 @@ public class FileDrop
     
     /**
      * Constructor with a default border and the option to recursively set drop targets.
-     * If your component is a <tt>java.awt.Container</tt>, then each of its children
+     * If your component is a java.awt.Container, then each of its children
      * components will also listen for drops, though only the parent will change borders.
      *
      * @param c Component on which files will be dropped.
      * @param recursive Recursively set children as drop targets.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -110,13 +110,13 @@ public class FileDrop
     /**
      * Constructor with a default border and debugging optionally turned on.
      * With Debugging turned on, more status messages will be displayed to
-     * <tt>out</tt>. A common way to use this constructor is with
-     * <tt>System.out</tt> or <tt>System.err</tt>. A <tt>null</tt> value for
-     * the parameter <tt>out</tt> will result in no debugging output.
+     * out. A common way to use this constructor is with
+     * System.out or System.err. A null value for
+     * the parameter out will result in no debugging output.
      *
      * @param out PrintStream to record debugging info or null for no debugging.
      * @param c Component on which files will be dropped.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -135,17 +135,17 @@ public class FileDrop
     /**
      * Constructor with a default border, debugging optionally turned on
      * and the option to recursively set drop targets.
-     * If your component is a <tt>java.awt.Container</tt>, then each of its children
+     * If your component is a java.awt.Container, then each of its children
      * components will also listen for drops, though only the parent will change borders.
      * With Debugging turned on, more status messages will be displayed to
-     * <tt>out</tt>. A common way to use this constructor is with
-     * <tt>System.out</tt> or <tt>System.err</tt>. A <tt>null</tt> value for
-     * the parameter <tt>out</tt> will result in no debugging output.
+     * out. A common way to use this constructor is with
+     * System.out or System.err. A null value for
+     * the parameter out will result in no debugging output.
      *
      * @param out PrintStream to record debugging info or null for no debugging.
      * @param c Component on which files will be dropped.
      * @param recursive Recursively set children as drop targets.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -167,8 +167,8 @@ public class FileDrop
      * Constructor with a specified border 
      *
      * @param c Component on which files will be dropped.
-     * @param dragBorder Border to use on <tt>JComponent</tt> when dragging occurs.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param dragBorder Border to use on JComponent when dragging occurs.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -188,13 +188,13 @@ public class FileDrop
     
     /**
      * Constructor with a specified border and the option to recursively set drop targets.
-     * If your component is a <tt>java.awt.Container</tt>, then each of its children
+     * If your component is a java.awt.Container, then each of its children
      * components will also listen for drops, though only the parent will change borders.
      *
      * @param c Component on which files will be dropped.
-     * @param dragBorder Border to use on <tt>JComponent</tt> when dragging occurs.
+     * @param dragBorder Border to use on JComponent when dragging occurs.
      * @param recursive Recursively set children as drop targets.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -215,14 +215,14 @@ public class FileDrop
     /**
      * Constructor with a specified border and debugging optionally turned on.
      * With Debugging turned on, more status messages will be displayed to
-     * <tt>out</tt>. A common way to use this constructor is with
-     * <tt>System.out</tt> or <tt>System.err</tt>. A <tt>null</tt> value for
-     * the parameter <tt>out</tt> will result in no debugging output.
+     * out. A common way to use this constructor is with
+     * System.out or System.err. A null value for
+     * the parameter out will result in no debugging output.
      *
      * @param out PrintStream to record debugging info or null for no debugging.
      * @param c Component on which files will be dropped.
-     * @param dragBorder Border to use on <tt>JComponent</tt> when dragging occurs.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param dragBorder Border to use on JComponent when dragging occurs.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -245,15 +245,15 @@ public class FileDrop
     /**
      * Full constructor with a specified border and debugging optionally turned on.
      * With Debugging turned on, more status messages will be displayed to
-     * <tt>out</tt>. A common way to use this constructor is with
-     * <tt>System.out</tt> or <tt>System.err</tt>. A <tt>null</tt> value for
-     * the parameter <tt>out</tt> will result in no debugging output.
+     * out. A common way to use this constructor is with
+     * System.out or System.err. A null value for
+     * the parameter out will result in no debugging output.
      *
      * @param out PrintStream to record debugging info or null for no debugging.
      * @param c Component on which files will be dropped.
-     * @param dragBorder Border to use on <tt>JComponent</tt> when dragging occurs.
+     * @param dragBorder Border to use on JComponent when dragging occurs.
      * @param recursive Recursively set children as drop targets.
-     * @param listener Listens for <tt>filesDropped</tt>.
+     * @param listener Listens for filesDropped.
      * @since 1.0
      */
     public FileDrop(
@@ -548,7 +548,7 @@ public class FileDrop
     }   // end isDragOk
     
     
-    /** Outputs <tt>message</tt> to <tt>out</tt> if it's not null. */
+    /** Outputs message to out if it's not null. */
     private static void log( java.io.PrintStream out, String message )
     {   // Log message if requested
         if( out != null )
@@ -609,7 +609,7 @@ public class FileDrop
     /**
      * Implement this inner interface to listen for when files are dropped. For example
      * your class declaration may begin like this:
-     * <code><pre>
+     *
      *      public class MyClass implements FileDrop.Listener
      *      ...
      *      public void filesDropped( java.io.File[] files )
@@ -617,7 +617,7 @@ public class FileDrop
      *          ...
      *      }   // end filesDropped
      *      ...
-     * </pre></code>
+     *
      *
      * @since 1.1
      */
@@ -626,7 +626,7 @@ public class FileDrop
         /**
          * This method is called when files have been successfully dropped.
          *
-         * @param files An array of <tt>File</tt>s that were dropped.
+         * @param files An array of Files that were dropped.
          * @since 1.0
          */
         public abstract void filesDropped( java.io.File[] files );
@@ -644,7 +644,7 @@ public class FileDrop
      * your {@link FileDrop.Listener} when files are dropped onto
      * a registered drop target.
      *
-     * <p>I'm releasing this code into the Public Domain. Enjoy.</p>
+     * I'm releasing this code into the Public Domain. Enjoy.
      * 
      * @author  Robert Harder
      * @author  rob@iharder.net
@@ -716,13 +716,13 @@ public class FileDrop
      *
      * The {@link java.awt.datatransfer.DataFlavor} associated with 
      * {@link TransferableObject} has the representation class
-     * <tt>net.iharder.dnd.TransferableObject.class</tt> and MIME type
-     * <tt>application/x-net.iharder.dnd.TransferableObject</tt>.
+     * net.iharder.dnd.TransferableObject.class and MIME type
+     * application/x-net.iharder.dnd.TransferableObject.
      * This data flavor is accessible via the static
      * {@link #DATA_FLAVOR} property.
      *
      *
-     * <p>I'm releasing this code into the Public Domain. Enjoy.</p>
+     * I'm releasing this code into the Public Domain. Enjoy.
      * 
      * @author  Robert Harder
      * @author  rob@iharder.net
@@ -732,7 +732,7 @@ public class FileDrop
     {
         /**
          * The MIME type for {@link #DATA_FLAVOR} is 
-         * <tt>application/x-net.iharder.dnd.TransferableObject</tt>.
+         * application/x-net.iharder.dnd.TransferableObject.
          *
          * @since 1.1
          */
@@ -742,9 +742,9 @@ public class FileDrop
         /**
          * The default {@link java.awt.datatransfer.DataFlavor} for
          * {@link TransferableObject} has the representation class
-         * <tt>net.iharder.dnd.TransferableObject.class</tt>
+         * net.iharder.dnd.TransferableObject.class
          * and the MIME type 
-         * <tt>application/x-net.iharder.dnd.TransferableObject</tt>.
+         * application/x-net.iharder.dnd.TransferableObject.
          *
          * @since 1.1
          */
@@ -764,7 +764,7 @@ public class FileDrop
          * Along with the {@link #DATA_FLAVOR} associated with this class,
          * this creates a custom data flavor with a representation class 
          * determined from <code>data.getClass()</code> and the MIME type
-         * <tt>application/x-net.iharder.dnd.TransferableObject</tt>.
+         * application/x-net.iharder.dnd.TransferableObject.
          *
          * @param data The data to transfer
          * @since 1.1
@@ -798,7 +798,7 @@ public class FileDrop
          * Along with the {@link #DATA_FLAVOR} associated with this class,
          * this creates a custom data flavor with a representation class <var>dataClass</var>
          * and the MIME type
-         * <tt>application/x-net.iharder.dnd.TransferableObject</tt>.
+         * application/x-net.iharder.dnd.TransferableObject.
          *
          * @see Fetcher
          * @param dataClass The {@link java.lang.Class} to use in the custom data flavor
@@ -812,7 +812,7 @@ public class FileDrop
 
         /**
          * Returns the custom {@link java.awt.datatransfer.DataFlavor} associated
-         * with the encapsulated object or <tt>null</tt> if the {@link Fetcher}
+         * with the encapsulated object or null if the {@link Fetcher}
          * constructor was used without passing a {@link java.lang.Class}.
          *
          * @return The custom data flavor for the encapsulated object
@@ -883,7 +883,7 @@ public class FileDrop
 
 
         /**
-         * Returns <tt>true</tt> if <var>flavor</var> is one of the supported
+         * Returns true if <var>flavor</var> is one of the supported
          * flavors. Flavors are supported using the <code>equals(...)</code> method.
          *
          * @param flavor The data flavor to check